Change code coverage to only support cobertura for now
