+++ JenkinsTests/CoberturaCoverageTests.swift
@@ -1,5 +1,5 @@
 //
-//  CoverageTests.swift
+//  CoberturaCoverageTests.swift
 //  Jenkins
 //
 //  Created by Patrick Butkiewicz on 11/13/16.
@@ -10,7 +10,7 @@ import Foundation
 import XCTest
 @testable import Jenkins
 
-class CoverageTests: XCTestCase {
+class CoberturaCoverageTests: XCTestCase {
     
     func testCoberturaCodeCoverageDepth2() {
         
@@ -25,7 +25,7 @@ class CoverageTests: XCTestCase {
                 return XCTFail("Failed mapping Coverage Report JSON")
         }
         
-        let c = CodeCoverageReport(json: json["results"] as! JSON)
+        let c = CoberturaCodeCoverageReport(json: json["results"] as! JSON)
         
         let validChildren = 0
         XCTAssert(c.childReports.count == validChildren, "Report has \(c.childReports.count) children, but should have \(validChildren)")
@@ -37,7 +37,8 @@ class CoverageTests: XCTestCase {
         XCTAssert(c.coverageElements.count == validCovElements, "Report has \(c.coverageElements.count) coverage elements but should have \(validCovElements)")
         
         let validLineRatio = 0.2
-        XCTAssert(c.lineRatio() == validLineRatio, "Ratio is \(c.lineRatio()) but should be \(validLineRatio)")
+        let actualLineRatio = c.ratio(of: .Lines)
+        XCTAssertEqualWithAccuracy(actualLineRatio, validLineRatio, accuracy: 0.05, "Ratio is \(actualLineRatio) but should be \(validLineRatio)")
     }
     
     func testCoberturaCodeCoverageDepth3() {
@@ -52,7 +53,7 @@ class CoverageTests: XCTestCase {
                 return XCTFail("Failed mapping Coverage Report JSON")
         }
         
-        let c = CodeCoverageReport(json: json["results"] as! JSON)
+        let c = CoberturaCodeCoverageReport(json: json["results"] as! JSON)
         
         let validChildren = 27
         XCTAssert(c.childReports.count == validChildren, "Report has \(c.childReports.count) children, but should have \(validChildren)")
@@ -63,8 +64,9 @@ class CoverageTests: XCTestCase {
         let validCovElements = 5
         XCTAssert(c.coverageElements.count == validCovElements, "Report has \(c.coverageElements.count) coverage elements but should have \(validCovElements)")
         
-        let validLineRatio = 0.20
-        XCTAssertEqualWithAccuracy(c.lineRatio(), validLineRatio, accuracy: 0.05, "Ratio is \(c.lineRatio()) but should be \(validLineRatio)")
+        let validLineRatio = 0.2
+        let actualLineRatio = c.ratio(of: .Lines)
+        XCTAssertEqualWithAccuracy(actualLineRatio, validLineRatio, accuracy: 0.05, "Ratio is \(actualLineRatio) but should be \(validLineRatio)")
     }
     
 }
